requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.129.0
h11==0.16.0
idna==3.11
kafka-python==2.3.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
redis==7.2.0
requests==2.32.5
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
